* master: (60 commits)
  Shrink sidebar social icons for a bit. (1.9em -> 1.5em)
  Tune icon sizes.
  Erase feed icon on the menubar.
  make icon-size tunable.
  Change texts for social links to large colored icons.
  update template.md
  content/README.md: on article creation.
  delete legacy create.sh
  content/create.py created.
  content/README.md: on auto-assignment of label and auto-post of preview link
  badge color 'darkgreen' -> green
  make tag_cloud a bit more compact.
  typo corrected.
  Hide feeds on sidebar.
  sidebar.html: SIDEBAR_HIDE_FEEDS option implemented.
  base.html: atom/rss feeds can be set by SHOW_{ATOM,RSS}_ON_MENU. defaults are true/false respectively, same as the original theme.
  Add RSS feed on Menu.
  Make taglist font a bit bolder.
  Horizontal alignment for tag-group list.
  Different colors for different tag-groups.
  ...

# Conflicts:
#	theme/voidy-bootstrap/static/css/voidybootstrap-custom.css

+++ tools/gotobranch.sh
@@ -1,5 +1,11 @@
 #!/bin/sh
 branch=${1:-"master"}
+git fetch
+if [ ! `git branch -a | sed 's/^[ \t]*//' | grep "^remotes/origin/$branch$"` ]; then
+    git branch -D $branch
+    echo "No branch $branch."
+    exit
+fi
 rm -rf content/*
 rm -rf theme/*
 git checkout -f $branch
